Minor refactoring of the files
